requirements.txt
requests>=2.31.0
colorama>=0.4.6
argparse>=1.4.0
pyyaml>=6.0
glob2>=0.7
pathlib>=1.0.1
subprocess32>=3.5.4; python_version < '3.0'
